+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - Removed attribute *+cdesc*. Replace by *+api(cdesc)*
 - Removed *f_module_line* and *c_module_line* from both statements and Typemap.
+- Remove *f_imports* from statements.
 
 ## v0.13.0 - 2023-08-23
 ### Added